+++ exts/cesium.omniverse/cesium/omniverse/ui/attributes/globe_anchor_attributes_widget.py
@@ -24,7 +24,9 @@ class CesiumGlobeAnchorAttributesWidget(SchemaPropertiesWidget):
                 CustomLayoutProperty("cesium:anchor:adjustOrientationForGlobeWhenMoving")
                 CustomLayoutProperty("cesium:anchor:detectTransformChanges")
             with CustomLayoutGroup("Global Positioning"):
-                CustomLayoutProperty("cesium:anchor:geographicCoordinates")
+                CustomLayoutProperty("cesium:anchor:latitude")
+                CustomLayoutProperty("cesium:anchor:longitude")
+                CustomLayoutProperty("cesium:anchor:height")
             with CustomLayoutGroup("Advanced Positioning", collapsed=True):
                 CustomLayoutProperty("cesium:anchor:position")
                 CustomLayoutProperty("cesium:anchor:rotation")