+++ src/plugins/CesiumUsdSchemas/src/CesiumUsdSchemas/tokens.h
@@ -51,14 +51,22 @@ struct CesiumTokensType {
     /// 
     /// CesiumGlobeAnchorAPI
     const TfToken cesiumAnchorDetectTransformChanges;
-    /// \brief "cesium:anchor:geographicCoordinates"
-    /// 
-    /// CesiumGlobeAnchorAPI
-    const TfToken cesiumAnchorGeographicCoordinates;
     /// \brief "cesium:anchor:georeferenceBinding"
     /// 
     /// CesiumGlobeAnchorAPI
     const TfToken cesiumAnchorGeoreferenceBinding;
+    /// \brief "cesium:anchor:height"
+    /// 
+    /// CesiumGlobeAnchorAPI
+    const TfToken cesiumAnchorHeight;
+    /// \brief "cesium:anchor:latitude"
+    /// 
+    /// CesiumGlobeAnchorAPI
+    const TfToken cesiumAnchorLatitude;
+    /// \brief "cesium:anchor:longitude"
+    /// 
+    /// CesiumGlobeAnchorAPI
+    const TfToken cesiumAnchorLongitude;
     /// \brief "cesium:anchor:position"
     /// 
     /// CesiumGlobeAnchorAPI